fix search bar, break out to another component
